requirements.txt
streamlit
easyocr
opencv-python-headless
pillow
numpy
pandas
pyspellchecker
torch
torchvision
transformers